+++ apps/block_scout_web/lib/block_scout_web/etherscan.ex
@@ -261,6 +261,12 @@ defmodule BlockScoutWeb.Etherscan do
     "result" => "21265524714464"
   }
 
+  @stats_ethsupplyexchange_example_value %{
+    "status" => "1",
+    "message" => "OK",
+    "result" => "101959776311500000000000000"
+  }
+
   @stats_ethsupply_example_value %{
     "status" => "1",
     "message" => "OK",
@@ -1772,9 +1778,35 @@ defmodule BlockScoutWeb.Etherscan do
     ]
   }
 
+  @stats_ethsupplyexchange_action %{
+    name: "ethsupplyexchange",
+    description: "Get total supply in Wei from exchange.",
+    required_params: [],
+    optional_params: [],
+    responses: [
+      %{
+        code: "200",
+        description: "successful operation",
+        example_value: Jason.encode!(@stats_ethsupplyexchange_example_value),
+        model: %{
+          name: "Result",
+          fields: %{
+            status: @status_type,
+            message: @message_type,
+            result: %{
+              type: "integer",
+              description: "The total supply.",
+              example: ~s("101959776311500000000000000")
+            }
+          }
+        }
+      }
+    ]
+  }
+
   @stats_ethsupply_action %{
     name: "ethsupply",
-    description: "Get total supply in Wei.",
+    description: "Get total supply in Wei from DB.",
     required_params: [],
     optional_params: [],
     responses: [
@@ -2302,6 +2334,7 @@ defmodule BlockScoutWeb.Etherscan do
     name: "stats",
     actions: [
       @stats_tokensupply_action,
+      @stats_ethsupplyexchange_action,
       @stats_ethsupply_action,
       @stats_ethprice_action
     ]