+++ apps/block_scout_web/lib/block_scout_web/views/api/rpc/stats_view.ex
@@ -7,6 +7,10 @@ defmodule BlockScoutWeb.API.RPC.StatsView do
     RPCView.render("show.json", data: Decimal.to_string(token_supply))
   end
 
+  def render("ethsupplyexchange.json", %{total_supply: total_supply}) do
+    RPCView.render("show.json", data: total_supply)
+  end
+
   def render("ethsupply.json", %{total_supply: total_supply}) do
     RPCView.render("show.json", data: total_supply)
   end